+++ packages/frontend/src/pages/Models.svelte
@@ -11,7 +11,7 @@ import TasksProgress from '/@/lib/progress/TasksProgress.svelte';
 import Card from '/@/lib/Card.svelte';
 import { onMount } from 'svelte';
 import ModelColumnSize from '../lib/table/model/ModelColumnSize.svelte';
-import ModelColumnCreation from '../lib/table/model/ModelColumnCreation.svelte';
+import ModelColumnAge from '../lib/table/model/ModelColumnAge.svelte';
 import ModelColumnActions from '../lib/table/model/ModelColumnActions.svelte';
 import Tab from '/@/lib/Tab.svelte';
 import Route from '/@/Route.svelte';
@@ -22,7 +22,7 @@ const columns: Column<ModelInfo>[] = [
   new Column<ModelInfo>('', { width: '40px', renderer: ModelColumnIcon }),
   new Column<ModelInfo>('Name', { width: '3fr', renderer: ModelColumnName }),
   new Column<ModelInfo>('Size', { width: '50px', renderer: ModelColumnSize }),
-  new Column<ModelInfo>('Creation', { width: '60px', renderer: ModelColumnCreation }),
+  new Column<ModelInfo>('Age', { width: '50px', renderer: ModelColumnAge }),
   new Column<ModelInfo>('', { width: '220px', align: 'right', renderer: ModelColumnLabels }),
   new Column<ModelInfo>('Actions', { align: 'right', width: '120px', renderer: ModelColumnActions }),
 ];
